+++ resources/views/account/addresses.blade.php
@@ -1,4 +1,4 @@
-<x-layouts::storefront :title="__('My Addresses')">
+<x-layouts::storefront :title="__('My Addresses')" :noindex="true">
     <x-account.layout>
         <livewire:storefront.address-book />
     </x-account.layout>