+++ resources/views/account/notifications.blade.php
@@ -1,4 +1,4 @@
-<x-layouts::storefront :title="__('Notifications')">
+<x-layouts::storefront :title="__('Notifications')" :noindex="true">
     <x-account.layout>
         <livewire:storefront.notifications-page />
     </x-account.layout>