+++ resources/views/account/orders-show.blade.php
@@ -1,4 +1,4 @@
-<x-layouts::storefront :title="__('Order detail')">
+<x-layouts::storefront :title="__('Order detail')" :noindex="true">
     <x-account.layout>
         <livewire:storefront.order-detail-page :order-ulid="$order" />
     </x-account.layout>