+++ resources/views/account/orders.blade.php
@@ -1,4 +1,4 @@
-<x-layouts::storefront :title="__('My Orders')">
+<x-layouts::storefront :title="__('My Orders')" :noindex="true">
     <x-account.layout>
         <livewire:storefront.order-history-page />
     </x-account.layout>