+++ resources/views/layouts/auth.blade.php
@@ -1,3 +1,3 @@
-<x-layouts::auth.simple :title="$title ?? null">
+<x-layouts::auth.simple :title="$title ?? null" :noindex="$noindex ?? false">
     {{ $slot }}
 </x-layouts::auth.simple>