+++ resources/views/orders/confirmation.blade.php
@@ -1,3 +1,3 @@
-<x-layouts::storefront :title="__('Order confirmed')">
+<x-layouts::storefront :title="__('Order confirmed')" :noindex="true">
     <livewire:storefront.order-confirmation-page :order-ulid="$order" />
 </x-layouts::storefront>