+++ resources/views/pages/login-phone.blade.php
@@ -1,3 +1,3 @@
-<x-layouts::auth :title="__('Log in with phone')">
+<x-layouts::auth :title="__('Log in with phone')" :noindex="true">
     <livewire:auth.phone-login />
 </x-layouts::auth>