+++ resources/views/wishlist/show.blade.php
@@ -1,4 +1,4 @@
-<x-layouts::storefront :title="__('My Wishlist')">
+<x-layouts::storefront :title="__('My Wishlist')" :noindex="true">
     <x-account.layout>
         <livewire:storefront.wishlist-page />
     </x-account.layout>