+++ routes/web.php
@@ -4,11 +4,15 @@ use App\Http\Controllers\Admin\SystemCacheController;
 use App\Http\Controllers\Auth\GoogleAuthController;
 use App\Http\Controllers\Storefront\AccountController;
 use App\Http\Controllers\Storefront\HomeController;
+use App\Http\Controllers\Storefront\RobotsController;
+use App\Http\Controllers\Storefront\SitemapController;
 use App\Http\Controllers\Storefront\StaticPageController;
 use App\Http\Controllers\Storefront\ThemeCssController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/theme.css', ThemeCssController::class)->name('theme.css');
+Route::get('/robots.txt', RobotsController::class)->name('robots.txt');
+Route::get('/sitemap.xml', SitemapController::class)->name('sitemap.xml');
 
 Route::get('/', [HomeController::class, 'show'])->name('home');